refactor(examples): extract server routing logic
